+++ src/app/action.ts
@@ -2,6 +2,7 @@
 
 import { ClientService } from "@/services/client"
 import { clientCreate, clientCreateSchema} from "@/drizzle/schema"
+import { revalidatePath } from "next/cache";
 
 const serviceMethods = new ClientService()
 
@@ -10,7 +11,8 @@ export async function clientCreateAction(data:clientCreate) {
         const parsedData = clientCreateSchema.parse(data)
 
         await serviceMethods.createClient(parsedData)
-        return {success:true,message:`Client ${parsedData.name} added successfully`}
+        revalidatePath('/'); // Revalidate the home page
+        return {success:true, message:`Client ${parsedData.name} added successfully`}
     } catch (err:any) {
         return {success:false, message: `Unable to add Client ${data.name}, due to ${err}`}
     }
