+++ src/components/dashboard/table.tsx
@@ -1,51 +1,491 @@
-import { clientSelect } from "@/drizzle/schema";
+"use client";
+
+import { cn } from "@/lib/utils";
 import {
-    Table,
-    TableBody,
-    TableCaption,
-    TableCell,
-    TableFooter,
-    TableHeader,
-    TableRow,
-    TableHead
-} from "../ui/table";
-import { FC } from "react";
-import { Button } from "../ui/button";
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
+import { Button } from "@/components/ui/button";
+import { Checkbox } from "@/components/ui/checkbox";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuGroup,
+  DropdownMenuItem,
+  DropdownMenuLabel,
+  DropdownMenuSeparator,
+} from "@/components/ui/dropdown-menu";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import {
+  Pagination,
+  PaginationContent,
+  PaginationItem,
+} from "@/components/ui/pagination";
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select";
+import {
+  Table as TableComponent,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from "@/components/ui/table";
+import {
+  ColumnDef,
+  ColumnFiltersState,
+  FilterFn,
+  PaginationState,
+  flexRender,
+  getCoreRowModel,
+  getFacetedUniqueValues,
+  getFilteredRowModel,
+  getPaginationRowModel,
+  useReactTable,
+} from "@tanstack/react-table";
+import {
+  ChevronFirstIcon,
+  ChevronLastIcon,
+  ChevronLeftIcon,
+  ChevronRightIcon,
+  CircleAlertIcon,
+  CircleXIcon,
+  EllipsisIcon,
+  FilterIcon,
+  ListFilterIcon,
+  PlusIcon,
+  TrashIcon,
+  MoveUpRight
+} from "lucide-react";
+import { clientSelect } from "@/drizzle/schema";
+import { AddClientDialog } from "@/components/dashboard/addClient";
+import { useEffect, useState, useRef } from "react";
 
-interface DashboardTableProps{
-    clients:clientSelect[]
+type Item = clientSelect;
+
+// Custom filter function for multi-column searching
+const multiColumnFilterFn: FilterFn<Item> = (row, columnId, filterValue) => {
+  const searchableRowContent =
+    `${row.original.name} ${row.original.email}`.toLowerCase();
+  const searchTerm = (filterValue ?? "").toLowerCase();
+  return searchableRowContent.includes(searchTerm);
+};
+
+// Add a helper function for text truncation
+const truncateText = (text: string, limit: number = 60) => {
+  if (text.length <= limit) return text;
+  return `${text.slice(0, limit)}...`;
+};
+
+const columns: ColumnDef<Item>[] = [
+  {
+    id: "select",
+    header: ({ table }) => (
+      <Checkbox
+        checked={
+          table.getIsAllPageRowsSelected() ||
+          (table.getIsSomePageRowsSelected() && "indeterminate")
+        }
+        onCheckedChange={(value) => table.toggleAllPageRowsSelected(!!value)}
+        aria-label="Select all"
+      />
+    ),
+    cell: ({ row }) => (
+      <Checkbox
+        checked={row.getIsSelected()}
+        onCheckedChange={(value) => row.toggleSelected(!!value)}
+        aria-label="Select row"
+      />
+    ),
+    size: 28,
+    enableSorting: false,
+  },
+  {
+    header: "Client Name",
+    accessorKey: "name",
+    cell: ({ row }) => (
+      <div className="font-medium truncate" title={row.getValue("name")}>
+        {truncateText(row.getValue("name"))}
+      </div>
+    ),
+    size: 180,
+    filterFn: multiColumnFilterFn,
+    enableSorting: false,
+  },
+  {
+    header: "Client Email",
+    accessorKey: "email",
+    size: 220,
+    enableSorting: false,
+    cell: ({ row }) => (
+      <div className="truncate" title={row.getValue("email")}>
+        {truncateText(row.getValue("email"))}
+      </div>
+    ),
+  },
+  {
+    header: "Client Address",
+    accessorKey: "address",
+    cell: ({ row }) => (
+      <div className="truncate" title={row.getValue("address")}>
+        {truncateText(row.getValue("address"))}
+      </div>
+    ),
+    size: 220,
+    enableSorting: false,
+  },
+  {
+    id: "actions",
+    header: () => <span className="sr-only">Actions</span>,
+    cell: ({ row }) => (
+      <Button variant="outline" className="items-center" size="sm">
+        view<MoveUpRight className="w-5 h-5" />
+      </Button>
+    ),
+    size: 60,
+    enableHiding: false,
+  },
+];
+
+
+interface TableProps {
+  clients: clientSelect[];
 }
 
-export const DashboardTable :FC<DashboardTableProps> = ({clients})=>{
-    return (
-        <Table className={"w-full bg-[#0C0A09]"}>
-            <TableHeader className="" >
-                <TableRow className="*:border-border [&>:not(:last-child)]:border-r" >
-                    <TableHead className="w-10">S.No</TableHead>
-                    <TableHead>Client Name</TableHead>
-                    <TableHead>Client Email</TableHead>
-                    <TableHead>Client Address</TableHead>
-                    <TableHead>Act</TableHead>
+export function Table({ clients }: TableProps) {
+  const id = useId();
+  const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>([]);
+  const [pagination, setPagination] = useState<PaginationState>({
+    pageIndex: 0,
+    pageSize: 5,
+  });
+  const inputRef = useRef<HTMLInputElement>(null);
+
+  const [data, setData] = useState<Item[]>(() => {
+    return [...clients].sort((a, b) => 
+      new Date(b.created_at).getTime() - new Date(a.created_at).getTime()
+    );
+  });
+
+  const handleDeleteRows = () => {
+    const selectedRows = table.getSelectedRowModel().rows;
+    const updatedData = data.filter(
+      (item) => !selectedRows.some((row) => row.original.id === item.id)
+    );
+    setData(updatedData);
+    table.resetRowSelection();
+  };
+
+  const table = useReactTable({
+    data,
+    columns,
+    getCoreRowModel: getCoreRowModel(),
+    getPaginationRowModel: getPaginationRowModel(),
+    onPaginationChange: setPagination,
+    onColumnFiltersChange: setColumnFilters,
+    getFilteredRowModel: getFilteredRowModel(),
+    getFacetedUniqueValues: getFacetedUniqueValues(),
+    state: {
+      pagination,
+      columnFilters,
+    },
+  });
+
+  return (
+    <div className="space-y-4">
+      {/* Filters */}
+      <div className="flex flex-wrap items-center justify-between gap-3">
+        <div className="flex items-center gap-3">
+          {/* Filter by name or email */}
+          <div className="relative">
+            <Input
+              id={`${id}-input`}
+              ref={inputRef}
+              className={cn(
+                "peer min-w-60 ps-9",
+                Boolean(table.getColumn("name")?.getFilterValue()) && "pe-9"
+              )}
+              value={
+                (table.getColumn("name")?.getFilterValue() ?? "") as string
+              }
+              onChange={(e) =>
+                table.getColumn("name")?.setFilterValue(e.target.value)
+              }
+              placeholder="Filter by name or email..."
+              type="text"
+              aria-label="Filter by name or email"
+            />
+            <div className="text-muted-foreground/80 pointer-events-none absolute inset-y-0 start-0 flex items-center justify-center ps-3 peer-disabled:opacity-50">
+              <ListFilterIcon size={16} aria-hidden="true" />
+            </div>
+            {Boolean(table.getColumn("name")?.getFilterValue()) && (
+              <button
+                className="text-muted-foreground/80 hover:text-foreground focus-visible:border-ring focus-visible:ring-ring/50 absolute inset-y-0 end-0 flex h-full w-9 items-center justify-center rounded-e-md transition-[color,box-shadow] outline-none focus:z-10 focus-visible:ring-[3px] disabled:pointer-events-none disabled:cursor-not-allowed disabled:opacity-50"
+                aria-label="Clear filter"
+                onClick={() => {
+                  table.getColumn("name")?.setFilterValue("");
+                  if (inputRef.current) {
+                    inputRef.current.focus();
+                  }
+                }}
+              >
+                <CircleXIcon size={16} aria-hidden="true" />
+              </button>
+            )}
+          </div>
+        </div>
+        <div className="flex items-center gap-3">
+          {/* Delete button */}
+          {table.getSelectedRowModel().rows.length > 0 && (
+            <AlertDialog>
+              <AlertDialogTrigger asChild>
+                <Button className="ml-auto" variant="outline">
+                  <TrashIcon
+                    className="-ms-1 opacity-60"
+                    size={16}
+                    aria-hidden="true"
+                  />
+                  Delete
+                  <span className="bg-background text-muted-foreground/70 -me-1 inline-flex h-5 max-h-full items-center rounded border px-1 font-[inherit] text-[0.625rem] font-medium">
+                    {table.getSelectedRowModel().rows.length}
+                  </span>
+                </Button>
+              </AlertDialogTrigger>
+              <AlertDialogContent>
+                <div className="flex flex-col gap-2 max-sm:items-center sm:flex-row sm:gap-4">
+                  <div
+                    className="flex size-9 shrink-0 items-center justify-center rounded-full border"
+                    aria-hidden="true"
+                  >
+                    <CircleAlertIcon className="opacity-80" size={16} />
+                  </div>
+                  <AlertDialogHeader>
+                    <AlertDialogTitle>
+                      Are you absolutely sure?
+                    </AlertDialogTitle>
+                    <AlertDialogDescription>
+                      This action cannot be undone. This will permanently delete{" "}
+                      {table.getSelectedRowModel().rows.length} selected{" "}
+                      {table.getSelectedRowModel().rows.length === 1
+                        ? "row"
+                        : "rows"}
+                      .
+                    </AlertDialogDescription>
+                  </AlertDialogHeader>
+                </div>
+                <AlertDialogFooter>
+                  <AlertDialogCancel>Cancel</AlertDialogCancel>
+                  <AlertDialogAction onClick={handleDeleteRows}>
+                    Delete
+                  </AlertDialogAction>
+                </AlertDialogFooter>
+              </AlertDialogContent>
+            </AlertDialog>
+          )}
+          {/* Add client button */}
+          <AddClientDialog>
+            <Button className="ml-auto" variant="outline">
+              <PlusIcon
+                className="-ms-1 opacity-60"
+                size={16}
+                aria-hidden="true"
+              />
+              Add Client
+            </Button>
+          </AddClientDialog>
+        </div>
+      </div>
+
+      {/* Table */}
+      <div className="bg-background overflow-hidden rounded-md border">
+        <TableComponent className="w-full table-fixed">
+          <TableHeader>
+            {table.getHeaderGroups().map((headerGroup) => (
+              <TableRow key={headerGroup.id} className="*:border-border [&>:not(:last-child)]:border-r hover:bg-transparent">
+                {headerGroup.headers.map((header) => (
+                  <TableHead
+                    key={header.id}
+                    style={{ width: `${header.getSize()}px` }}
+                    className="h-11"
+                  >
+                    {header.isPlaceholder ? null : flexRender(
+                      header.column.columnDef.header,
+                      header.getContext()
+                    )}
+                  </TableHead>
+                ))}
+              </TableRow>
+            ))}
+          </TableHeader>
+          <TableBody>
+            {table.getRowModel().rows?.length ? (
+              table.getRowModel().rows.map((row) => (
+                <TableRow
+                  key={row.id}
+                  data-state={row.getIsSelected() && "selected"}
+                  className="*:border-border [&>:not(:last-child)]:border-r"
+                >
+                  {row.getVisibleCells().map((cell) => (
+                    <TableCell 
+                      key={cell.id} 
+                      className={cn(
+                        "last:py-0",
+                        cell.column.id === "name" && "font-medium"
+                      )}
+                    >
+                      {flexRender(
+                        cell.column.columnDef.cell,
+                        cell.getContext()
+                      )}
+                    </TableCell>
+                  ))}
                 </TableRow>
-            </TableHeader>
-            <TableBody  className="[&_td:first-child]:rounded-l-lg " >
-                {clients.map((client,index)=>{
-                    return (
-                        <TableRow key={client.id}  className="*:border-border  [&>:not(:last-child)]:border-r"
-                         >
-                            <TableCell className="p-4" >{index+1}</TableCell>
-                            <TableCell className="font-medium" >{client.name}</TableCell>
-                            <TableCell>{client.email}</TableCell>
-                            <TableCell>{client.address}</TableCell>
-                            <TableCell>
-                                <Button variant="outline"size="sm">
-                                    view
-                                </Button>
-                            </TableCell>
-                        </TableRow>
-                    )
-                })}
-            </TableBody>
-        </Table>
-    )
+              ))
+            ) : (
+              <TableRow>
+                <TableCell
+                  colSpan={columns.length}
+                  className="h-24 text-center"
+                >
+                  No results.
+                </TableCell>
+              </TableRow>
+            )}
+          </TableBody>
+        </TableComponent>
+      </div>
+
+      {/* Pagination */}
+      <div className="flex items-center justify-between gap-8">
+        {/* Results per page */}
+        <div className="flex items-center gap-3">
+          <Label htmlFor={id} className="max-sm:sr-only">
+            Rows per page
+          </Label>
+          <Select
+            value={table.getState().pagination.pageSize.toString()}
+            onValueChange={(value) => {
+              table.setPageSize(Number(value));
+            }}
+          >
+            <SelectTrigger id={id} className="w-fit whitespace-nowrap">
+              <SelectValue placeholder="Select number of results" />
+            </SelectTrigger>
+            <SelectContent className="[&_*[role=option]]:ps-2 [&_*[role=option]]:pe-8 [&_*[role=option]>span]:start-auto [&_*[role=option]>span]:end-2">
+              {[5, 10, 25, 50].map((pageSize) => (
+                <SelectItem key={pageSize} value={pageSize.toString()}>
+                  {pageSize}
+                </SelectItem>
+              ))}
+            </SelectContent>
+          </Select>
+        </div>
+        {/* Page number information */}
+        <div className="text-muted-foreground flex grow justify-end text-sm whitespace-nowrap">
+          <p
+            className="text-muted-foreground text-sm whitespace-nowrap"
+            aria-live="polite"
+          >
+            <span className="text-foreground">
+              {table.getState().pagination.pageIndex *
+                table.getState().pagination.pageSize +
+                1}
+              -
+              {Math.min(
+                Math.max(
+                  table.getState().pagination.pageIndex *
+                    table.getState().pagination.pageSize +
+                    table.getState().pagination.pageSize,
+                  0
+                ),
+                table.getRowCount()
+              )}
+            </span>{" "}
+            of{" "}
+            <span className="text-foreground">
+              {table.getRowCount().toString()}
+            </span>
+          </p>
+        </div>
+
+        {/* Pagination buttons */}
+        <div>
+          <Pagination>
+            <PaginationContent>
+              {/* First page button */}
+              <PaginationItem>
+                <Button
+                  size="icon"
+                  variant="outline"
+                  className="disabled:pointer-events-none disabled:opacity-50"
+                  onClick={() => table.firstPage()}
+                  disabled={!table.getCanPreviousPage()}
+                  aria-label="Go to first page"
+                >
+                  <ChevronFirstIcon size={16} aria-hidden="true" />
+                </Button>
+              </PaginationItem>
+              {/* Previous page button */}
+              <PaginationItem>
+                <Button
+                  size="icon"
+                  variant="outline"
+                  className="disabled:pointer-events-none disabled:opacity-50"
+                  onClick={() => table.previousPage()}
+                  disabled={!table.getCanPreviousPage()}
+                  aria-label="Go to previous page"
+                >
+                  <ChevronLeftIcon size={16} aria-hidden="true" />
+                </Button>
+              </PaginationItem>
+              {/* Next page button */}
+              <PaginationItem>
+                <Button
+                  size="icon"
+                  variant="outline"
+                  className="disabled:pointer-events-none disabled:opacity-50"
+                  onClick={() => table.nextPage()}
+                  disabled={!table.getCanNextPage()}
+                  aria-label="Go to next page"
+                >
+                  <ChevronRightIcon size={16} aria-hidden="true" />
+                </Button>
+              </PaginationItem>
+              {/* Last page button */}
+              <PaginationItem>
+                <Button
+                  size="icon"
+                  variant="outline"
+                  className="disabled:pointer-events-none disabled:opacity-50"
+                  onClick={() => table.lastPage()}
+                  disabled={!table.getCanNextPage()}
+                  aria-label="Go to last page"
+                >
+                  <ChevronLastIcon size={16} aria-hidden="true" />
+                </Button>
+              </PaginationItem>
+            </PaginationContent>
+          </Pagination>
+        </div>
+      </div>
+    </div>
+  );
 }
+function useId() {
+  const [id] = useState(() => `table-${Math.random().toString(36).slice(2, 9)}`);
+  return id;
+}
+
